Change error handling
This commit effectively reverts commit 752cbfe9. I'm sure this won't be
the last of my wishy washy design either. The previous simplification
came about because Entropy weighting method needs to normalize the input
matrix before weighting the criteria. This change led to a particular
issue with the way potential errors were handled, specifically
Sum::normalize()? potentially passing a NormalizationError, but the
weighting function returning a WeightingError. Having separate error
types is good for naming and keeping things logically together, but can
cause quite a bit of duplicate code, which is why they were combined.

This latest attempt to make something meaningful, all of the common
error variants (which are admittedly all errors currently), are put into
their own error enum and then included in the superset for Weighting,
Normalization, and Ranking which allows for the possibility of better
scoping potential errors for the modules, while allowing coercion of
common errors such as in the Entropy weighting method where the matrix
needs normalized.