Refactor code and docs to use CriteriaType enum
The criteria type are always represented as a cost or a profit and that
representation typically is in the form of -1 and 1 respectively.
This refactor takes all of the checking out of the normalization
functions and allows them to take an array slice of enum CriteriaType
which is guarenteed to be only a Cost or Profit once the functions are
ready to use it for their calculations.