Fix documentation inconsistencies
Many of the methods mathemetical descriptions are not standardized. E.g.,
sometimes arrays are n x m, other times m x n. Likewise, when summing over
columns, the count is over the number of columns when it should be over the
number of rows.

With a lack of a standard for myself, I'll go with what wikipedia has for
matrices and say a matrix is m rows by n columns