Give criteria types function parameters better name
Lazily, the `CriteriaTypes` function parameters were just named `types`
which isn't nearly as descriptive as it could be and requires the user
to already expect the kinds of parameters each normalization or ranking
function must have. Naming the `CriteriaTypes` parameter to
`criteria_types` immediately tells the user what the parameter is for.