Fix order of matrix validation checks
While it may not truly matter, it just seems strange to assign the size
of the array to variables, just to return if those sizes are zero. The
compiler may take care of this kinds of optimization anyway, but for
piece of mind, this change checks for an empty matrix before assigning
variables.