Change `scale_columns` to `apply_column_weights`
Naming is hard...I've struggled with this function for so long.
Hopefully this name is a little more clear on what the function is doing
and sticks around for a while. I'm sure there is still a better name out
there somewhere. I only hope that one day it makes it way to me and I
can give this function the name it deserves.