Re-export traits directly for ergonomics
Currently, a user would need to add something like
`use crate::normalization::Normalize`. From the perspective of the code
this approach is good and keeps all methods that belong to the
`Normalize` trait contained in the `normalization` module, but this
looks pretty gross and redundant when incorporating the modules in the
code. This cleans up the used traits a little.