Update rust edition
I suppose it's fine to keep things the way they are if they are working,
but I believe there is more value keeping the tooling up-to-date. This
forces us to embrace change the new releases of software, tooling,
libraries, etc. We also get in the habbit of using the most recent
editions and get the bug fixes and optimizations that come along with
the updates. What I do hope is that rust version and edition changes
don't explode like C++ has and the rust community stays sane about the
roadmap and progress.