Remove used of `Vec<f64>`
Ultimately, I'd like for this library to have the option to use no_std.
This change removes the reliance on the standard libraries `Vec` which
requires an allocator thats typically not suitable for no_std. This
changes uses DVector directly so we can rely on the way nalgebra handles
dynamic allocation in their no_std support.